requirements.txt
pyqt6
opencv-python
numpy
depthai
pytest
Pillow
